+++ src/LandingPage.tsx
@@ -10,7 +10,7 @@ import {
   Stack,
 } from "@mui/material";
 
-import Header from "./Header";
+import Header from "./UI/Header";
 import MuiDropzone from "./UI/MuiDropzone";
 import VFEList from "./VFEList";
